donut_service/requirements.txt
# Flask web framework
Flask>=2.3.0
flask-cors>=4.0.0

# Donut model dependencies (CPU-only for development)
transformers>=4.30.0
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cpu
sentencepiece>=0.1.99
Pillow>=10.0.0
